added 'caching', fixed submission timestamp, added full year support
the flask backend now tracks when it has last pulled data from advent
of code in order to not overflow the website.
the submission timestamp was computed incorrectly, the from the
timestamp YYYY:DD:05:00 was subtracted but had to be YYYY:DD:05:00
which is fixed now.
if this website gets run during another time of the year it will
'clamp' the date to 25th of december of the previous year, or this
year if it is still december.