reformatted to use datetime fields/functions and not strf and strp
previously the day, month and year of a datetime object were gotten by
calling strf, additionally as a result the day month and year were
stored as a string.
changed this to store this information as an integer.
datetimes are now created using the datetime() object instead of strp.
reformatted some lines (strings are now formatted using format strings
instead of appending and renamed an index value from int to i as int
is an internal function).