assured compliance with the guidelines of aoc automation
on the subreddit wiki there are some guidelines on creating
automations for advent of code posted by a maintainer of aoc.
these guidelines state that requests must include a link to the code
repo and the email of the maintainers in the user-agent header.
beside this requests must be throttled and responses cached where
possible (e.g. puzzles do not change, so request once and serve forever).